+++ src/parser/parser.ts
@@ -1,3 +1,4 @@
+import { EOF_TOKEN, type Token, TokenType, isToken } from 'src/core/token';
 /* eslint-disable no-cond-assign */
 import {
   AllColumnsAsterisk,
@@ -13,7 +14,6 @@ import {
   Statement,
   TokenNode,
 } from './ast';
-import { EOF_TOKEN, type Token, TokenType, isToken } from './token';
 
 /**
  * A simple parser that creates a very rudimentary syntax tree.
